tests/wpt/web-platform-tests/old-tests/submission/Opera/script_scheduling/051.html

<!DOCTYPE html>
<html><head>
	<title> scheduler: interaction of parsing and script execution - script added through DOM</title>
	<script src="/resources/testharness.js"></script>
	<script src="/resources/testharnessreport.js"></script>
	<script src="testlib/testlib.js"></script>
</head>
<body>

	<div id="log">FAILED (This TC requires JavaScript enabled)</div>

	<script>log('inline script #1');
		testlib.addScript('', { src: 'scripts/count-script-tags.js' }, document.body, true);
		log('end script #1');
	</script>

	<script type="text/javascript">
	log( 'inline script #2' );
	var t = async_test()

	function test() {
		assert_any(assert_array_equals, eventOrder, [['inline script #1', 'end script #1', 'script tags in DOM: 5', 'inline script #2'],
                                                             ['inline script #1', 'end script #1', 'inline script #2', 'script tags in DOM: 6']]);
		t.done();
        }
	onload = function(){setTimeout(t.step_func(test), 100); }
	</script>

</body></html>
